% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shedz.R
\name{bc_delin_basin}
\alias{bc_delin_basin}
\title{Delineate a watershed from Canadian Digital Elevation Data within British Columbia}
\usage{
bc_delin_basin(
  pour_point,
  xmin,
  xmax,
  ymin,
  ymax,
  convergence = 1.1,
  minslope = 0.1
)
}
\arguments{
\item{pour_point}{Geo-referenced sf point object, make sure pour point lies exactly within channel of interest.}

\item{xmin}{Estimated minimum X coordinate in map units defining extent of basin to delineate.}

\item{xmax}{Estimated maximum X coordinate in map units defining extent of basin to delineate.}

\item{ymin}{Estimated minimum Y coordinate in map units defining extent of basin to delineate.}

\item{ymax}{Estimated maximum Y coordinate in map units defining extent of basin to delineate.}

\item{convergence}{Convergence parameter for SAGA MFD algorithm.}

\item{minslope}{Minimum slope parameter for SAGA fill DEM algorithm.}
}
\value{
A list with the delineated basin as a sf polygon, and the CDED raster object for the extent provided.
}
\description{
This function requires a pour point and extent to delineate a watershed
from digital elevation data derived from CDED, obtained using \link[bcmaps]{cded}.
Flow tracing is done using \link\link{RSAGA} and multiple flow directions (MFD). Other
transformations are done using gdal (gdalwarp,gdal_traslate & gdal_polygonize) and the \link\link{raster} package.
}
